patch 8.2.3355: MS-Windows: compiler warning for 64-32 bit conversion
Problem:    MS-Windows: compiler warning for 64-32 bit conversion.
Solution:   Add type casts.

+++ src/memline.c
@@ -4000,7 +4000,7 @@ ml_flush_line(buf_T *buf)
 		dp->db_txt_start -= extra;
 #if defined(FEAT_BYTEOFF) && defined(FEAT_PROP_POPUP)
 		if (buf->b_has_textprop)
-		    old_prop_len = old_len - STRLEN(new_line) - 1;
+		    old_prop_len = old_len - (int)STRLEN(new_line) - 1;
 #endif
 
 		// copy new line into the data block
@@ -4012,7 +4012,7 @@ ml_flush_line(buf_T *buf)
 		{
 		    // Do not count the size of any text properties.
 		    extra += old_prop_len;
-		    extra -= new_len - STRLEN(new_line) - 1;
+		    extra -= new_len - (int)STRLEN(new_line) - 1;
 		}
 		if (extra != 0)
 		    ml_updatechunk(buf, lnum, (long)extra, ML_CHNK_UPDLINE);
